Crate v_utils

Re-exports

pub extern crate v_utils_macros as macros;

Re-exports

pub use other::*;

Modules

io

other

prelude

trades

utils

Macros

clientside

Fallback when xdg is not available - logs to stdout

define_str_enum

elog

Macro for logging to both stderr and tracing debug Usage: elog!(“message”) or elog!(“format {}”, arg)

fmt_with_width

log

Macro for logging to both stdout and tracing info Usage: log!(“message”) or log!(“format {}”, arg)

xdg_cache_dir

Will create $XDG_CACHE_HOME/<crate_name>/$subpath/ (“” for no subpath; subpath is a DIR)

xdg_cache_file

Will create $XDG_CACHE_HOME/<crate_name>/ and return the path to the file specified in $subpath

xdg_config_dir

Will create $XDG_CONFIG_HOME/<crate_name>/$subpath/ (“” for no subpath; subpath is a DIR)

xdg_config_file

Will create $XDG_CONFIG_HOME/<crate_name>/ and return the path to the file specified in $subpath

xdg_data_dir

Will create $XDG_DATA_HOME/<crate_name>/$subpath/ (“” for no subpath; subpath is a DIR)

xdg_data_file

Will create $XDG_DATA_HOME/<crate_name>/ and return the path to the file specified in $subpath

xdg_runtime_dir

Will create $XDG_RUNTIME_DIR/<crate_name>/$subpath/ (“” for no subpath; subpath is a DIR)

xdg_runtime_file

Will create $XDG_RUNTIME_DIR/<crate_name>/ and return the path to the file specified in $subpath

xdg_state_dir

Will create $XDG_STATE_HOME/<crate_name>/$subpath/ (“” for no subpath; subpath is a DIR)

xdg_state_file

Will create $XDG_STATE_HOME/<crate_name>/ and return the path to the file specified in $subpath