v_utils

Module prelude

Re-exports

pub use std::fmt::Write as _;

pub use std::str::FromStr as _;

pub use eyre::OptionExt as _;

pub use eyre::WrapErr as _;

pub use crate::io::ExpandedPath;

pub use crate::other::*;

pub use crate::trades::*;

Modules

clientsideDeprecated

instrument

Attach a span to a std::future::Future.

libsideDeprecated

Macros

bail

Return early with an error.

clientside

Fallback when xdg is not available - logs to stdout

debug

Constructs an event at the debug level.

error

Constructs an event at the error level.

eyre

Construct an ad-hoc error from a string.

info

Constructs an event at the info level.

json

Construct a serde_json::Value from a JSON literal.

trace

Constructs an event at the trace level.

warn

Constructs an event at the warn level.

Structs

Arc

A thread-safe reference-counting pointer. ‘Arc’ stands for ‘Atomically Reference Counted’.

BTreeMap

An ordered map based on a B-Tree.

BTreeSet

An ordered set based on a B-Tree.

Empty

An empty field.

HashMap

A hash map implemented with quadratic probing and SIMD lookup.

HashSet

A hash set implemented as a HashMap where the value is ().

Mutex

A mutual exclusion primitive useful for protecting shared data

Pin

A pointer which pins its pointee in place.

Report

The core error reporting type of the library, a wrapper around a dynamic error reporting type.

RwLock

A reader-writer lock

Span

A handle representing a span, with the capability to enter the span if it exists.

VecDeque

A double-ended queue implemented with a growable ring buffer.

Enums

Value

Represents any valid JSON value.

Traits

Deserialize

A data structure that can be deserialized from any data format supported by Serde.

DeserializeOwned

A data structure that can be deserialized without borrowing any data from the deserializer.

Deserializer

A data format that can deserialize any data structure supported by Serde.

Serialize

A data structure that can be serialized into any data format supported by Serde.

Serializer

A data format that can serialize any data structure supported by Serde.

Functions

join_all

Creates a future which represents a collection of the outputs of the futures given.

Type Aliases

Result

type alias for Result<T, Report>

Attribute Macros

instrument

Instruments a function to create and enter a tracing span every time the function is called.

Derive Macros

Deserialize

Error

Serialize