v_utils::prelude

Module instrument

Attach a span to a std::future::Future.

Structs

Instrumented

A Future that has been instrumented with a tracing Span.

WithDispatch

A Future that has been instrumented with a tracing Subscriber.

Traits

Instrument

Attaches spans to a std::future::Future.

WithSubscriber

Extension trait allowing futures to be instrumented with a tracing Subscriber.