v_utils::trades

Module klines

Structs

Close

Unlike Kline, timestamp signifies the end of the period, not the start. As another difference - Vec<Close> can have uneven spacing between measured points.

Kline

Standard candlestick data unit. Can only ever be full, - if an exchange returns partial data for an ongoing candle, or if trading/exchange is down leading to the associated data being cut, the Kline object is NOT created.

Ohlc

Functions

mock_p_to_ohlc

take a price-series, and imagine that entries are constantly spaced

p_to_ohlc