v_utils::utils::snapshots

Function snapshot_plot_orders

pub fn snapshot_plot_orders<T: Into<f64> + Copy>(
    prices: &[T],
    orders: &[(usize, Option<T>)],
) -> String

Panics

if ordinals on orders are outside of prices or not ascending.

Blocker

Until better fonts, distinctions between price formats, multiple order lines at a time & order types, actual timeframes, are all extremely problematic; so their implementation is postponed.

Architecture

Uses SnapshotP to build the plot, for finer control use it instead.